// src/types/analytics.ts
export enum AnalyticsEventType {
  subscriptionPlansView = "subscription_plans_view",
  subscribeNowClick = "subscribe_now_click",
  homepageView = "homepage_view",
  thumbnailView = "thumbnail_view",
  thumbnailClick = "thumbnail_click",
  loginTry = "login_try",
  loginSuccess = "login_sucess",
  forgetPasswordTry = "forget_password_try",
  forgetPasswordMailSend = "forget_password_mail_send",
  viewIntroPages = "view_intro_pages",
  signupTry = "signup_try",
  signupMailSent = "started_new_plan",
  selectProfile = "select_profile",
  editProfile = "edit_profile",
  addProfile = "add_profile",
  viewHomePage = "view_home_page",
  clickThumbnail = "click_thumbnail",
  clickSeeAll = "click_seeall",
  clickPlayTopBanner = "click_play_top_banner",
  clickOnPlayIcon = "click_on_play_icon",
  clickOnPlayButton = "click_on_play_button",
  clickDownload = "click_download",
  addMyList = "add_mylisty",
  clickShareMovie = "click_share_movie",
  clickLikeMovie = "click_like_movie",
  clickEpisode = "click_episode",
  clickMoreLike = "click_more_like",
  clickPause = "click_pause",
  clickPlay = "click_play",
  clickNext = "click_next",
  clickPrevious = "click_previous",
  clickShowSubtitle = "click_show_subtitle",
  showMyListMovie = "show_mylist_movie",
  showLiveTvList = "show_live_tv_list",
  changeLanguage = "change_language",
  openProfileManager = "open_profile_manager",
  openSupport = "open_support",
  openFaq = "open_FAQ",
  openTermsCondition = "open_terms_condition",
  notificationOn = "notification_on",
  notificationOff = "notification_off",
  logout = "logout",
  openHistoryMovie = "open_history_movie",
  playHistoryVideos = "play_history_videos",
  searchByQuery = "seach_by_query",
  openSearchMovie = "open_search_movie",
  updateSubscriptionByMovie = "update_subscription_by_movie",
  updateSubscriptionByProfile = "started_renew_plan",
  showDownload = "show_download",
  adClicked = "ad_clicked",
  adWatched = "ad_watched",
  adSkipped = "ad_skipped",
}

export interface AnalyticsPayload {
  eventType: string;
  movieId?: string;
  episodeId?: string;
  adId?: string;
}